requirements.txt
tensorflow>=2.12.0
opencv-python
matplotlib
seaborn
scikit-learn
Pillow
tqdm
